pyproject.toml
[tool.poetry]
name = "dl"
version = "0.1.0"
description = ""
authors = ["Daniel Correia <example@email.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dl"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.1"
pillow = "^10.2.0"
matplotlib = "^3.8.3"
torchvision = "^0.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
